3)db: update dockerfile and run the command cargo sqlx prepare to use the macro query_as! and buil the container

+++ bot/migrations/001_init.sql
@@ -4,5 +4,5 @@ CREATE TABLE apy_snapshots (
     protocol TEXT NOT NULL,
     network TEXT NOT NULL,
     asset TEXT NOT NULL,
-    apy NUMERIC
+    apy DOUBLE PRECISION
 );
